requirements_test.txt
--requirement requirements.txt

autopep8==2.0.4
coverage==7.3.0
nose-exclude==0.5.0

django-jenkins==0.110.0
django-nose==1.4.7
django-pytest==0.2.0
docutils==0.20.1
factory-boy==3.3.0
flake8==6.1.0

flower==2.0.1
HTTPretty==1.1.4
mock==5.1.0
nose==1.3.7
rednose==1.3.0
pep8==1.7.1

pyflakes==3.1.0
pycodestyle==2.11.0
# clonedigger==1.1.0
shortuuid==1.0.11
Pygments==2.16.1
PyHamcrest==2.0.4
pytest==7.4.0
pytest-cov==4.1.0
sphinx==7.2.5
# sphinx-rtd-theme==1.3.0
sure==2.0.1
#urllib3==1.21.1

Jinja2==3.1.2
MarkupSafe==2.1.3
PyYAML==6.0.1
certifi==2023.7.22
chardet==5.2.0
# gtextfsm==0.2.1
lxml==4.9.3
ncclient==0.6.13
ntc-templates==3.5.0
pyparsing==3.1.1
pyserial==3.5
scp==0.14.5
selectors2==2.0.2
terminal==0.4.0
transitions==0.9.0
urllib3==2.0.4
yamlordereddictloader==0.4.0